Main Changes:   - Fixed a few bugs   - Added pictures for view, edit, delte, etc. table actions

+++ lifttracker/view/lifts-view.php
@@ -19,7 +19,7 @@
         <thead>
           <tr>
             <th scope="col"><a href="#">Name</a></th>
-            <th scope="col"><a href="#">Training Weight</a></th>
+            <th scope="col"><a href="#">Weight</a></th>
             <th scope="col"></th>
             <th scope="col"></th>
             <th scope="col"></th>
@@ -29,12 +29,11 @@
         <tbody>
   <?php foreach ($lifts as $lift) { ?>
           <tr>
-            <td><?php echo(htmlspecialchars($lift->getTitle())); ?></td>
+            <td><a href="../main/index.php?controller=lift&action=liftview&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>"><?php echo(htmlspecialchars($lift->getTitle())); ?></a></td>
             <td><?php echo(htmlspecialchars($lift->getTrainingWeight())); ?></td>
-            <td><a href="../main/index.php?controller=lift&action=liftview&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>">View</a></td>
-            <td><a href="../main/index.php?controller=lift&action=liftedit&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>">Edit</a></td>
-            <td><a href="../main/index.php?controller=lift&action=liftdelete&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>">Delete</a></td>
-            <td><a href="../main/index.php?controller=attempt&action=attemptadd&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>">Add Attempt</a></td>
+            <td><a href="../main/index.php?controller=lift&action=liftedit&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>"><img src="../images/edit.png" alt="Edit" /></a></td>
+            <td><a data-toggle="confirmation" href="../main/index.php?controller=lift&action=liftdelete&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>"><img src="../images/delete.png" alt="Delete" /></a></td>
+            <td><a href="../main/index.php?controller=attempt&action=attemptadd&liftid=<?php echo(htmlspecialchars($lift->getId())); ?>"><img src="../images/add.png" alt="Add Attempt" /></a></td>
           </tr>
   <?php } ?>
         </tbody>
